Possible frontend design
During last nights coding event, we created a vague sketch of how the website may look like. This is the suggested design which I'd like to discuss and whether or not we should to go forward with it.
If we settle with this design I would like to start a discussion about main components and whether or not the components styling should be inside the component or in a separate .css file, I would suggest the former.